+++ files/gitlab-ctl-commands/upgrade.rb
@@ -54,14 +54,15 @@ add_command 'upgrade', 'Run migrations after a package upgrade', 1 do |cmd_name|
   end
 
   log 'Shutting down all GitLab services except those needed for migrations'
-  get_all_services.each do |sv_name|
+  MIGRATION_SERVICES = %w(postgresql redis geo-postgresql gitaly).freeze
+  (get_all_services - MIGRATION_SERVICES).each do |sv_name|
     run_sv_command_for_service('stop', sv_name)
   end
 
-  MIGRATION_SERVICES = %w(postgresql redis geo-postgresql gitaly).freeze
+  log 'Ensuring the required services are running'
   MIGRATION_SERVICES.each do |sv_name|
     # If the service is disabled, e.g. because we are using an external
-    # Postgres server, then this command is a no-op.
+    # Postgres server, or it is already running, then this command is a no-op.
     run_sv_command_for_service('start', sv_name)
   end